This is a major update, the bulk of which is the seperation of the underlying window system and rendering context creation code from the main body of the 3d engine, and the implementation of GTK+/gtkglext module that can be chosen instead of SDL.
Major changes:
- removed SDL code from 3dengfx.cpp and placed it in init_sdl.cpp
- placed all the event handling and wm-related SDL code of fxwt.cpp in
  conditionally compiled blocks.
- added GTK gl initalization and the assorted gtk widgetry to be used
  in GTK programs as well as the usual event-handling through fxwt
  events. (proper event handling is in progress, no mouse events yet,
  and idle functions are ignored at this point; instead the gfx update
  functions are called on idle).
- removed the exception mechanism that was used before as well as the
  logger functionality which is superseded by the error/warning/info
  variadic functions. Previously exception throwing functions are
  changed to return success/failure and log automatically.
- excluded the fxwt widget stuff from build. The fxwt widgets
  development is posponed at this point, GTK is much more powerful
  anyway and there is no pressing need for in-scene 3D widgetry.

Bug fixes:
- fixed horrible bug in config_parser.c which could (and did at some
  point) trash malloc data structures by calling realloc on an already
  freed buffer if it tried to parse a second config file.

Preperations for 0.2 release:
- Added directories install (for installation and configuration
  scripts/programs) and examples (for example programs).
- Removed the old 3dengfx-config bash script and created a C program
  that does the same job while also taking under consideration the
  various configuration parameters in 3dengfx_config.h

+++ src/common/config_parser.c
@@ -55,7 +55,7 @@ void SetParserState(enum ParserState state, int value) {
 int LoadConfigFile(const char *fname) {
 	FILE *fp;
 	int fsize;
-	char *temp;
+	char *temp, *line;
 	
 	if(!fname) return -1;	
 	if(!(fp = fopen(fname, "r"))) return -1;
@@ -68,9 +68,9 @@ int LoadConfigFile(const char *fname) {
 	config_file = temp;
 	
 	cfgptr = config_file;
-	temp = malloc(max_line_len + 1);
-	while(fgets(temp, max_line_len, fp)) {
-		char *ptr = temp;
+	if(!(line = malloc(max_line_len + 1))) return -1;
+	while(fgets(line, max_line_len, fp)) {
+		char *ptr = line;
 		
 		if(*ptr == '\n') continue;	/* kill empty lines, they irritate the parser */
 		
@@ -81,7 +81,7 @@ int LoadConfigFile(const char *fname) {
 			ptr++;
 		}
 		
-		if(*ptr == sym_comment && ptr != temp) {
+		if(*ptr == sym_comment && ptr != line) {
 			*cfgptr++ = '\n';
 		}
 	}
@@ -90,7 +90,7 @@ int LoadConfigFile(const char *fname) {
 	
 	memset(&cfg_opt, 0, sizeof(struct ConfigOption));
 	cfgptr = config_file;
-	free(temp);
+	free(line);
 	return 0;
 }
 
@@ -137,4 +137,5 @@ void DestroyConfigParser() {
 	if(cfg_opt.str_value) free(cfg_opt.str_value);
 	if(cfg_opt.option) free(cfg_opt.option);
 	if(config_file) free(config_file);
+	config_file = 0;
 }
